pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "jose>=1.0.0",
    "openai>=1.97.0",
    "passlib>=1.7.4",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "reportlab>=4.4.2",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
